### 2024-10-15 #7353 @editorialbot
[PRE REVIEW]: ExpFamilyPCA.jl: A Julia Package for Exponential Family Principal Component Analysis
**Submitting author:** @FlyingWorkshop (<a href="http://orcid.org/0009-0001-3978-9462">Logan Bhamidipaty</a>)
**Repository:** https://github.com/sisl/ExpFamilyPCA.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
