### 2024-10-22 #7353 @FlyingWorkshop
Hi @Irnv, thanks for agreeing to be our editor!

Here are some reviewers I'd recommend:
1. [Guillaume Dalle](https://reviewers.joss.theoj.org/reviewers/2816) - active Julia contributor and familiar with statistics, optimization, and autodiff (we use symbolic diff, but the expertise should transfer)
2. [Manuel Stapper](https://reviewers.joss.theoj.org/reviewers/3149) - knows Julia and has expertise with computational statistics and count data (our `PoissonEPCA` model is designed to work with count data)
3. [Oliver Dunbar](https://reviewers.joss.theoj.org/reviewers/3343) - knows Julia + math and statistics
4.  [Charles Dufour](https://reviewers.joss.theoj.org/reviewers/2647) - knows Julia + stats
5. [Miroslav Kratochvíl](https://reviewers.joss.theoj.org/reviewers/4089) - knows Julia and is familiar with dimensionality reduction

Let me know if this helps and if there's anything else I can do to assist!
