### 2024-10-23 #7353 @lrnv
@ManuelStapper, @dufourc1, @exaexa, @gdalle, @lsandig, would some of you guys agree to review this submission for JOSS ? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
