### 2024-10-24 #7403 @lrnv
👋🏼 @ManuelStapper, @gdalle, @dufourc1, @FlyingWorkshop : this is the review thread for the paper. All of our communications will happen here from now on, so that each of us can read what each other writes.

@ManuelStapper, @gdalle, @dufourc1, as reviewers, your first step is to create a checklist for yourself, to guide you through the review, by commenting

on this thread. These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use @editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping me (@lrnv) if you have any questions/concerns, if something is unclear, or if you need more info about the process :)
