+++ paper.bib
@@ -3,7 +3,8 @@
   author={Collins, Michael and Dasgupta, Sanjoy and Schapire, Robert E},
   journal={Advances in Neural Information Processing Systems},
   volume={14},
-  year={2001}
+  year={2001},
+  doi={10.7551/mitpress/1120.003.0084}
 }
 
 @book{elements,
@@ -11,7 +12,8 @@
   author={Hastie, Trevor and Tibshirani, Robert and Friedman, Jerome H and Friedman, Jerome H},
   volume={2},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-0-387-84858-7}
 }
 
 @article{composition,
@@ -105,7 +107,7 @@ abstract = {Small signals may contain important information. Mass spectra of che
   pages={211--246},
   year={2001},
   publisher={Springer},
-  doi={https://doi.org/10.1023/A:1010896012157}
+  doi={10.1023/A:1010896012157}
 }
 
 @article{forster,
@@ -121,22 +123,6 @@ url = {https://www.sciencedirect.com/science/article/pii/S0022000001917982},
 author = {Jürgen Forster and Manfred K. Warmuth},
 abstract = {In the literature a number of relative loss bounds have been shown for on-line learning algorithms. Here the relative loss is the total loss of the on-line algorithm in all trials minus the total loss of the best comparator that is chosen off-line. However, for many applications instantaneous loss bounds are more interesting where the learner first sees a batch of examples and then uses these examples to make a prediction on a new instance. We show relative expected instantaneous loss bounds for the case when the examples are i.i.d. with an unknown distribution. We bound the expected loss of the algorithm on the last example minus the expected loss of the best comparator on a random example. In particular, we study linear regression and density estimation problems and show how the leave-one-out loss can be used to prove instantaneous loss bounds for these cases. For linear regression we use an algorithm that is similar to a new on-line learning algorithm developed by Vovk. Recently a large number of relative total loss bounds have been shown that have the form O(lnT), where T is the number of trials/examples. Standard conversions of on-line algorithms to batch algorithms result in relative expected instantaneous loss bounds of the form O(lnTT). Our methods lead to O(1T) upper bounds. In many cases we give tight lower bounds.}
 }
-
-@article{faithful,
- ISSN = {00359254, 14679876},
- URL = {http://www.jstor.org/stable/2347385},
- abstract = {An analysis of data on the duration times and waiting times for eruptions from the {O}ld {F}aithful {G}eyser reveals an interesting time series structure. A tentative physical model, derived from Rinehart, is outlined and a corresponding first-order Markov chain examined. It is shown that a second-order models is necessary to explain the observed correlations in the data. A curious clustering effect is apparent in the autocorrelation function when plotted over a large range of lags. Similar patterns are observed in simulations from the fitted second-order model.},
- author = {A. Azzalini and A. W. Bowman},
- journal = {Journal of the Royal Statistical Society. Series C (Applied Statistics)},
- number = {3},
- pages = {357--365},
- publisher = {[Royal Statistical Society, Oxford University Press]},
- title = {A Look at Some Data on the Old Faithful Geyser},
- urldate = {2024-10-09},
- volume = {39},
- year = {1990}
-}
-
 
 @article{gans,
 	author = {Gan, Guojun and Valdez, Emiliano A.},
@@ -171,6 +157,7 @@ abstract = {In the literature a number of relative loss bounds have been shown f
   year      =  1989,
   address   = "Philadelphia, PA",
[…]
