+++ paper.bib
@@ -157,7 +157,6 @@ abstract = {In the literature a number of relative loss bounds have been shown f
   year      =  1989,
   address   = "Philadelphia, PA",
   language  = "en"
-  doi = {
 }
 
 @article{optim, doi = {10.21105/joss.00615}, url = {https://doi.org/10.21105/joss.00615}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {24}, pages = {615}, author = {Patrick K. Mogensen and Asbjørn N. Riseth}, title = {Optim: A mathematical optimization package for {J}ulia}, journal = {Journal of Open Source Software} }
