### 2024-10-29 eeb7e1a
added JOSS badge to readme
