### 2024-11-04 cfcc98a
added contribution guidelines, internal links, and expanded documentation index
