+++ docs/src/index.md
@@ -1,6 +1,8 @@
 # ExpFamilyPCA.jl Documentation
 
-**ExpFamilyPCA.jl** is a Julia package for performing [exponential principal component analysis (EPCA)](https://papers.nips.cc/paper_files/paper/2001/hash/f410588e48dc83f2822a880a68f78923-Abstract.html). ExpFamilyPCA.jl supports custom objectives and includes fast implementations for several common distributions.
+**ExpFamilyPCA.jl** is a Julia package for performing exponential principal component analysis (EPCA) [EPCA](@cite). EPCA generalizes PCA to accommodate data from any exponential family distribution, making it more suitable for fields where these data types are common, such as geochemistry, marketing, genomics, political science, and machine learning [composition, elements](@cite).
+
+ExpFamilyPCA.jl the first EPCA package in Julia and the first in any language to support EPCA for multiple distributions and custom objectives.
 
 ## Installation
 
@@ -34,18 +36,24 @@ Y_reconstructed = decompress(poisson_epca, Y_compressed)
 
 | Distribution         | `ExpFamilyPCA.jl`                 | Objective                                                         | Link Function `` g(\theta) ``                                    |
 |----------------------|-----------------------------------|-------------------------------------------------------------------|------------------------------------------------------------------|
-| Bernoulli            | `BernoulliEPCA`                   | `` \log(1 + e^{\theta - 2x\theta}) ``                             | `` \frac{e^\theta}{1 + e^\theta} ``                              |
-| Binomial             | `BinomialEPCA`                    | `` n \log(1 + e^\theta) - x\theta ``                              | `` \frac{ne^\theta}{1 + e^\theta} ``                             |
-| Continuous Bernoulli | `ContinuousBernoulliEPCA`         | `` \log\left(\frac{e^\theta - 1}{\theta}\right) - x\theta ``      | `` \frac{\theta - 1}{\theta} + \frac{1}{e^\theta - 1} ``         |
-| Gamma¹    | `GammaEPCA` or `ItakuraSaitoEPCA` | `` -\log(-\theta) - x\theta ``                                   | `` -\frac{1}{\theta} ``                                          | 
-| Gaussian² | `GaussianEPCA` or `NormalEPCA`    | `` \frac{1}{2}(x - \theta)^2 ``                                   | `` \theta ``                                                     |
-| Negative Binomial    | `NegativeBinomialEPCA`            | `` -r \log(1 - e^\theta) - x\theta ``                             | `` \frac{-re^\theta}{e^\theta - 1} ``                            |
-| Pareto               | `ParetoEPCA`                      | `` -\log(-1 - \theta) + \theta \log m - x\theta ``                | `` \log m - \frac{1}{\theta + 1} ``                              |
-| Poisson³  | `PoissonEPCA`                     | `` e^\theta - x\theta ``                                          | `` e^\theta ``                                                   |
-| Weibull              | `WeibullEPCA`                     | `` -\log(-\theta) - x\theta ``                                    | `` -\frac{1}{\theta} ``                                          |
+| Bernoulli            | [`BernoulliEPCA`](./constructors/bernoulli.md)                   | `` \log(1 + e^{\theta - 2x\theta}) ``                             | `` \frac{e^\theta}{1 + e^\theta} ``                              |
+| Binomial             | [`BinomialEPCA`](./constructors/binomial.md)                    | `` n \log(1 + e^\theta) - x\theta ``                              | `` \frac{ne^\theta}{1 + e^\theta} ``                             |
[…]
