### 2024-11-04 #7403 @FlyingWorkshop
@ManuelStapper, thanks so much for the review! I've updated the documentation based on your feedback:

I expanded the documentation landing page to explain the **statement of need** for ExpFamilyPCA.jl. It largely matches the paper and explains potential fields that the package could be used in (e.g., geochemistry, marketing, genomics). Let me know if there are other details you'd recommend including.

**Installation instructions** and **example usage** can be found on the [GitHub README](https://github.com/sisl/ExpFamilyPCA.jl) and on the [documentation landing page](https://github.com/sisl/ExpFamilyPCA.jl).

**Functionality documentation** can also be found on the documentation website (parametric distribution constructors are found in the sidebar and [other API atoms are documented here](https://sisl.github.io/ExpFamilyPCA.jl/stable/api/)). I also added links to the table on the landing page for improved navigability.

Lastly, I've also added a section on **community guidelines** at the bottom of the documentation landing page.
