+++ dev/refs.bib
@@ -16,6 +16,26 @@
   publisher={Springer}
 }
 
+@book{elements,
+  title={The {E}lements of {S}tatistical {L}earning: {D}ata {M}ining, {I}nference, and {P}rediction},
+  author={Hastie, Trevor and Tibshirani, Robert and Friedman, Jerome H and Friedman, Jerome H},
+  volume={2},
+  year={2009},
+  publisher={Springer},
+  doi={10.1007/978-0-387-84858-7}
+}
+
+@article{composition,
+  title={Compositional data analysis},
+  author={Greenacre, Michael},
+  journal={Annual Review of Statistics and its Application},
+  volume={8},
+  number={1},
+  pages={271--299},
+  year={2021},
+  publisher={Annual Reviews}
+}
+
 @misc{dispatch,
   author       = {Stefan Karpinski},
   title        = {The Unreasonable Effectiveness of Multiple Dispatch},
