### 2024-11-10 sisl/ExpFamilyPCA.jl#21 @gdalle
JOSS review
Hi, and congrats on your JOSS submission! I'm one of your reviewers and I'll be writing my remarks below as the review progresses. You can find my checklist [here](https://github.com/openjournals/joss-reviews/issues/7403#issuecomment-2434630470).

## Paper

- The paper is [longer than 1000 words](https://github.com/openjournals/joss-reviews/issues/7403#issuecomment-2434575986), which is the [recommended limit](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). However I don't think shortening it would be helpful, because it does contain useful mathematical background.
- There are [missing and invalid DOIs](https://github.com/openjournals/joss-reviews/issues/7403#issuecomment-2434576033) among the references.
- L15: The right citation for the Julia language is usually the [SIAM Review paper](https://julialang.org/research/), not the Arxiv preprint you cited.
- L30-32: How do the following ingredients come into play inside your package? This is not at all explained in the paper.
  - symbolic differentiation
  - optimization
  - numerically stable computation
- L30-32: Why do you use symbolic differentiation (with Symbolics.jl) and not algorithmic differentiation (e.g. with Zygote.jl or Enzyme.jl)?
- L32: The citation you picked for "numerically stable computation" does not seem related to Julia but to R? In fact, I would assume that this third ingredient is also widely available in other languages as well.
- L33: What kind of benchmarks allow you to state that "`ExpFamilyPCA.jl` delivers speed"?
- L54: It would be useful to define the exponential family in general (as well as its natural and mean parameter) before diving into the details of the link function.
- L61: The link called "appendix" points to a documentation page, which may become out of date if you change the structure of your docs. The same goes for other external links in the paper.
- L81: Is it standard to introduce regularization around a $\mu_0$, or did you suggest it yourselves? How do you pick $\mu_0$ in practice?
- L88: Which figure do you recreate? Can you give more details on what these "belief profiles" represent?
- L88: Is the figure recreation reproducible?
- L91: "PCA struggles even with 10 basis" is missing the word "components".
- L104: It would help to clarify the types of the objects that `fit!`, `compress` and `decompress` work with.
