### 2024-11-10 sisl/ExpFamilyPCA.jl#21 @gdalle
## Code

- You may want to run CI on the latest Julia version (called 1 in the GitHub action you use) in addition to 1.10.
- It is useful to display what percentage of your code is covered by the test suite. At the moment, the Codecov part of the test workflow is [failing](https://github.com/sisl/ExpFamilyPCA.jl/actions/runs/11671794695/job/32498988141) because you did not specify a token. See [this tutorial](https://discourse.julialang.org/t/psa-new-version-of-codecov-action-requires-additional-setup/109857) to set it up properly.
- You may want to turn your example from the README into a [Documenter doctest](https://github.com/thchr/TestReadme.jl) to make sure it doesn't get outdated.
