### 2024-11-10 sisl/ExpFamilyPCA.jl#21 @gdalle
## Documentation

- - You may want to turn your example from the README into a [Documenter doctest](https://github.com/thchr/TestReadme.jl) to make sure it doesn't get outdated.
