### 2024-11-10 #7403 @gdalle
I have completed my review, overall this is really exciting work! Still there are few things to iron out before this is publication ready: you can find my suggestions in https://github.com/sisl/ExpFamilyPCA.jl/issues/21. I'm at the disposal of the authors if anything seems unclear.
