### 2024-11-10 6da9cbd
Improve type-stability
