+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"constructors/poisson/#Poisson-EPCA","page":"Poisson","title":"Poisson EPCA","text":"","category":"section"},{"location":"constructors/poisson/","page":"Poisson","title":"Poisson","text":"Name PoissonEPCA\nG(theta) e^theta\ng(theta) e^theta\nmu Space[1] positive\nTheta Space real\nAppropriate Data count, probability","category":"page"},{"location":"constructors/poisson/","page":"Poisson","title":"Poisson","text":"Poisson EPCA minimizes the generalized KL divergence making it well-suited for compressing probability profiles. Poisson EPCA has also been used in reinforcement learning to solve partially observed Markov decision processes (POMDPs) with belief compression (Roy et al., 2005). ","category":"page"},{"location":"constructors/poisson/","page":"Poisson","title":"Poisson","text":"[1]: mu space refers to the space of valid regularization parameters, not to the expectation parameter space.","category":"page"},{"location":"constructors/poisson/#Documentation","page":"Poisson","title":"Documentation","text":"","category":"section"},{"location":"constructors/poisson/","page":"Poisson","title":"Poisson","text":"PoissonEPCA","category":"page"},{"location":"constructors/poisson/#ExpFamilyPCA.PoissonEPCA","page":"Poisson","title":"ExpFamilyPCA.PoissonEPCA","text":"PoissonEPCA(indim::Integer, outdim::Integer; options::Options = Options())\n\nPoisson EPCA.\n\nArguments\n\nindim::Integer: Dimension of the input space.\noutdim::Integer: Dimension of the latent (output) space.\noptions::Options: Optional parameters.\n\nReturns\n\nepca: An EPCA subtype for the Poisson distribution.\n\n\n\n\n\n","category":"function"},{"location":"math/references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"math/references/","page":"References","title":"References","text":"Azoury, K. S. and Warmuth, M. K. (2001). Relative Loss Bounds for On-Line Density Estimation with the Exponential Family of Distributions. Machine learning 43, 211–246.\n\n\n\nBoyd, S. and Vandenberghe, L. (2004). Convex Optimization (Cambridge University Press).\n\n\n\nBregman, L. (1967). The relaxation method of finding the common point of convex sets and its application to the solution of problems in convex programming. USSR Computational Mathematics and Mathematical Physics 7, 200–217.\n\n\n\nCollins, M.; Dasgupta, S. and Schapire, R. E. (2001). A Generalization of Principal Components Analysis to the Exponential Family. Advances in Neural Information Processing Systems 14.\n\n\n\nForster, J. and Warmuth, M. K. (2002). Relative Expected Instantaneous Loss Bounds. Journal of Computer and System Sciences 64, 76–102.\n\n\n\nGowda, S.; Ma, Y.; Cheli, A.; Gwóźzdź, M.; Shah, V. B.; Edelman, A. and Rackauckas, C. (2022). High-Performance Symbolic-Numerics via Multiple Dispatch. ACM Commun. Comput. Algebra 55, 92–96.\n\n\n\nGreenacre, M. (2021). Compositional data analysis. Annual Review of Statistics and its Application 8, 271–299.\n\n\n\nHastie, T.; Tibshirani, R.; Friedman, J. H. and Friedman, J. H. (2009). The Elements of Statistical Learning: Data Mining, Inference, and Prediction. Vol. 2 (Springer).\n\n\n\nItakura, F. and Saito, S. (1968). Analysis Synthesis Telephony Based on the Maximum Likelihood Meth
[…]
