+++ paper.bib
@@ -135,15 +135,17 @@ abstract = {In the literature a number of relative loss bounds have been shown f
 }
 
 
-@misc{Julia,
-      title={Julia: A Fast Dynamic Language for Technical Computing}, 
-      author={Jeff Bezanson and Stefan Karpinski and Viral B. Shah and Alan Edelman},
-      year={2012},
-      eprint={1209.5145},
-      archivePrefix={arXiv},
-      primaryClass={cs.PL},
-      doi={10.48550/arXiv.1209.5145},
-      url={https://doi.org/10.48550/arXiv.1209.5145}
+@article{Julia,
+    title={Julia: A fresh approach to numerical computing},
+    author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+    journal={SIAM {R}eview},
+    volume={59},
+    number={1},
+    pages={65--98},
+    year={2017},
+    publisher={SIAM},
+    doi={10.1137/141000671},
+    url={https://epubs.siam.org/doi/10.1137/141000671}
 }
 
 @BOOK{GLM,
