### 2024-11-18 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
Thank you again for all the incredibly detailed feedback @gdalle! Working on addressing the feedback as I go.

- [x] L15: Verify the citation for the Julia language; it should be the SIAM Review paper, not the Arxiv preprint.
- [x] L30-32: Explain how the following ingredients are integrated into your package: _Added a footnote of explanation in the paper_
  - [x] Symbolic differentiation
  - [x] Optimization
  - [x] Numerically stable computation
- [x] L30-32: Justify the use of symbolic differentiation (with Symbolics.jl) instead of algorithmic differentiation (e.g., Zygote.jl or Enzyme.jl). _Also addressed in the new footnote_
- [x] L32: Review the citation for "numerically stable computation"; it appears to reference R instead of Julia and might be more broadly applicable to other languages. _Also addressed in the new footnote_
- [ ] L33: Provide benchmarks to substantiate the claim that "ExpFamilyPCA.jl delivers speed."
- [ ] L54: Define the exponential family in general, including its natural and mean parameters, before discussing the link function.
- [ ] L61: Update the "appendix" link to avoid pointing to documentation that may become outdated; ensure other external links are similarly reviewed.
- [ ] L81: Clarify if introducing regularization around a specific  $\mu_0 $
  is standard practice or your own contribution. Explain how $\mu_0$ is chosen in practice.
- [ ] L88: Provide more details about the figure recreation, including:
  - [ ] Which figure is being recreated
  - [ ] What the "belief profiles" represent
  - [ ] Whether the figure recreation is reproducible
- [ ] L91: Add the missing word "components" to the sentence "PCA struggles even with 10 basis."
- [ ] L104: Clarify the object types that `fit!`, `compress`, and `decompress` work with.
