### 2024-11-18 c45df45
tweak footnote citation
