### 2024-11-19 9303105
updated regularization section of paper
