+++ paper.bib
@@ -174,7 +174,7 @@ issn = {0041-5553},
 doi = {10.1016/0041-5553(67)90040-7},
 url = {https://www.sciencedirect.com/science/article/pii/0041555367900407},
 author = {L.M. Bregman},
-abstract = {IN this paper we consider an iterative method of finding the common point of convex sets. This method can be regarded as a generalization of the methods discussed in [1–4]. Apart from problems which can be reduced to finding some point of the intersection of convex sets, the method considered can be applied to the approximate solution of problems in linear and convex programming.}
+abstract = {In this paper we consider an iterative method of finding the common point of convex sets. This method can be regarded as a generalization of the methods discussed in [1–4]. Apart from problems which can be reduced to finding some point of the intersection of convex sets, the method considered can be applied to the approximate solution of problems in linear and convex programming.}
 }
 
 @misc{logexp,
