### 2024-11-20 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
Answering these questions, as I have time. Will continue to update this post as I answer more questions.

> * You should probably run CI on the latest Julia version (called 1 in the GitHub action you use) in addition to 1.10.
> * You may want to turn your example from the index into a [README test](https://github.com/thchr/TestReadme.jl) to make sure it doesn't get outdated.
> * It is useful to display what percentage of your code is covered by the test suite. At the moment, the Codecov part of the test workflow is [failing](https://github.com/sisl/ExpFamilyPCA.jl/actions/runs/11671794695/job/32498988141) because you did not specify a token. See [this tutorial](https://discourse.julialang.org/t/psa-new-version-of-codecov-action-requires-additional-setup/109857) to set it up properly.

Agree to all the above, working on these.

> * Did you profile the code to see where it spends most of its time and avoid classic performance pitfalls (type instability for example)? If you care about optimal speed, you can take a look at the [Julia performance tips](https://docs.julialang.org/en/v1/manual/performance-tips/) for numerous ways to accelerate your code.
> * Are there any benchmarks of your package against competitors?

No. Besides traditional PCA, there are no other EPCA implementations in Julia, so not sure if benchmarking would make sense for most of the distributions. That said, I did do some testing with MultivariateStats.jl's implementation of traditional PCA which is faster than our implementation of Gaussian EPCA. I haven't looked at their source code, but I suspect it's because they use the closed form solution for PCA whereas we use the same general iterative optimization procedure for Gaussian EPCA that we use for all EPCA objectives. I suspect it would be very hard (and not entirely the package's focus) to implement a faster version of PCA than MultivariateStats.jl's.

> * For optimization, you seem to be using the default algorithm of Optim.jl, which is the zero-order Nelder-Mead algorithm. Is there any reason not to pick a first- or second-order method? With automatic differentiatin you can get at least gradients basically for free.

I did some crude benchmarking using `BenchmarkTools` while designing the optimization pipeline. Surprisingly, I found that higher order methods performed roughly the same as or much slower than NM.

> * The documentation on [EPCA objectives](https://sisl.github.io/ExpFamilyPCA.jl/stable/math/objectives/) mentions 7 ways to construt the EPCA object, but the `src/constructors` folder only shows 4. Any reason why those four get special treatment?

There are in fact many ways to derive the EPCA objectives, more than the 7 ways I listed in the documentation. The 4 I ended up picking where the ones that I believed would be most useful and efficient in practice. `EPCA1` through `EPCA4` each represent an 'archetypal' specification of the EPCA objective. For example, `EPCA1` represents the EPCA objective using $F$ and $g$. However, we can derive $g$ from $G$, so we can equally-well specify the EPCA objective from $F$ and $G$, but this specification is marginally slower than the former because it requires symbolic differentiation. Thus, $F$ and $g$ is a better 'archetypal' struct than $F$ and $G$. We can make similar arguments for the other numbered EPCA structs. In short, each numbered EPCA struct is a way to specify the EPCA objective that is quick in practice. The other methods housed in each file are just more more involved ways to specify the same archetype.

Let me know if this explanation makes sense. Happy to clarify.

> * It seems like CompressedBeliefMDPs.jl functionality is not useful for the majority of potential users, so maybe it could be a [package extension](https://pkgdocs.julialang.org/v1/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions)) instead of a hard dependency?

Agree, looking into this.

> * [Different results for two consecutive calls to `fit!` on the same data #22](https://github.com/sisl/ExpFamilyPCA.jl/issues/22)

Will try to add a test for this. Also saw your previous comment on changing `similar` to `copy`. That was a great catch and your analysis was very helpful; it's highly appreciated. Thank you.
