### 2024-11-20 sisl/ExpFamilyPCA.jl#21 @gdalle
> No. Besides traditional PCA, there are no other EPCA implementations in Julia, so not sure if benchmarking would make sense for most of the distributions. That said, I did do some testing with MultivariateStats.jl's implementation of traditional PCA which is faster than our implementation of Gaussian EPCA. I haven't looked at their source code, but I suspect it's because they use the closed form solution for PCA whereas we use the same general iterative optimization procedure for Gaussian EPCA that we use for all EPCA objectives. I suspect it would be very hard (and not entirely the package's focus) to implement a faster version of PCA than MultivariateStats.jl's.

Fair enough.

> I did some crude benchmarking using BenchmarkTools while designing the optimization pipeline. Surprisingly, I found that higher order methods performed roughly the same as or much slower than NM.

That is rather surprising. I think it might have been tied to the way you compute gradients or Hessians. It is definitely something I can help with if you want to collaborate, but it is not a barrier to paper acceptance by any stretch.

> In short, each numbered EPCA struct is a way to specify the EPCA objective that is quick in practice.

Maybe you could add a summary in the docs of which construction schemes (among the 7 you list) are fast and which are slightly slower?
