### 2024-11-29 #7403 @dufourc1
I finished my checklist.  Interesting work ! The manuscript looks good and is nice to read; my only other question (on the use of `Symbolics.jl`)  has already been addressed in https://github.com/sisl/ExpFamilyPCA.jl/issues/21. (Apologies for the delay)
