### 2024-11-30 #7403 @ManuelStapper
I also finished reviewing, really interesting work!
In addition to the comments of the other reviewers, I have two questions:
- In the paragraph on the Bregman Divergences, maybe replace “Linear approximation about q” by “Linear Taylor approximation around q evaluated at p”?
- Regarding the Poisson example: Did I understand correctly, that the data we apply the EPCA to does not necessarily have the be integer-valued if we apply the Poisson-EPCA? What is the reason for using the Poisson here instead of Exponential or Gamma?
