### 2024-12-02 #7403 @FlyingWorkshop
@ManuelStapper, great question! Normally you would expect Poisson EPCA to work best on Poisson-distributed data (i.e., count data); however, the math works out so that the Poisson EPCA objective is equivalent to the (generalized) KL-divergence (which is commonly used to measure a "statistical distance" between two probability distributions), see [this derivation in the documentation](https://sisl.github.io/ExpFamilyPCA.jl/stable/math/appendix/poisson/). This means that Poisson EPCA is more appropriate for compressing probability distributions than say the Gamma or Exponential EPCA.
