### 2024-12-06 feaa1ce
added definition of exp family
