### 2024-12-06 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
@gdalle, thanks again for all the thorough feedback! I've just finished responding to the bullets.
