### 2024-12-06 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
Regarding the README test: this is a great idea, but the snippets currently use `sample_from_poisson` which is a generic helper that is not implemented, so cannot test the method. I believe our current test suite does cover similar constructions as the README though (see [here](https://github.com/sisl/ExpFamilyPCA.jl/blob/96d9fc8fc0597e8fa19dc186dcfa0ebde117c959/test/utils.jl#L48)).

Regarding codecov, I don't have the admin privileges in the SISL GitHub org needed to set up codecov.
