### 2024-12-06 e19f93c
update documentation and paper
