### 2024-12-06 #7403 @FlyingWorkshop
I've finished responding to @gdalle's feedback. Thank you again for your thoroughness and suggestions, I think they really increased the quality and clarify of the paper/codebase.
