+++ versions.js
@@ -1,7 +1,8 @@
 var DOC_VERSIONS = [
   "stable",
+  "v2.0",
   "v1.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.1.2";
+var DOCUMENTER_NEWEST = "v2.0.0";
 var DOCUMENTER_STABLE = "stable";
