### 2024-12-10 #7403 @lrnv
Yes the version will be bump before publication that's planned
