### 2024-12-11 #7403 @lrnv
@gdalle Even if you did not check up every box, due to the discussion sin your issue & PR i'll consider your review finished, please manifest yourself otherwise.

@FlyingWorkshop Can you please take care of the few items of the post-review checklist below and then report to me versions/dois, so that we can move on to acceptance ?
