### 2024-12-11 #7403 @gdalle
I actually meant to take a second look because I did not validate that every one of my remarks had been addressed, and the package authors closed the issue I opened without my approval. I'll try to get around to it today
