### 2024-12-11 sisl/ExpFamilyPCA.jl#21 @gdalle
> Regarding the README test: this is a great idea, but the snippets currently use sample_from_poisson which is a generic helper that is not implemented, so cannot test the method.

Regardless of whether the README is tested during CI, it seems weird to put code in a code example that people can't run. Can you maybe replace `sample_from_poisson(...)` with `rand(Distributions.Poisson(lambda), ...)`?

> Regarding codecov, I don't have the admin privileges in the SISL GitHub org needed to set up codecov.

It's a shame, codecov is very useful to figure out if some parts of your code are untested. Can you ask one of the admins to set it up.

Since we don't have codecov yet, I looked at the tests to understand what they cover. Why do you run a full test on some distributions and only a smoke test on others?
