### 2024-12-11 sisl/ExpFamilyPCA.jl#21 @gdalle
I don't think you answered any of my [remarks on the documentation](https://github.com/sisl/ExpFamilyPCA.jl/issues/21#issuecomment-2466651354)?
