### 2024-12-12 9fab4da
improved documentation based on JOSS review
