+++ docs/src/math/bregman.md
@@ -1,12 +1,10 @@
 # Bregman Divergences
 
-Bregman divergences [Bregman](@cite) play a central role in the probabilistic framework of EPCA. They generalize the concept of distance between two points but do not necessarily satisfy the properties of a traditional metric (such as symmetry or the triangle inequality). Instead, Bregman divergences provide a flexible way to measure differences between data points, making them useful in applications like clustering, optimization, and information theory.
-
-Understanding Bregman divergences is essential for EPCA because they link the exponential family of probability distributions to loss functions used in optimization, allowing us to generalize PCA to non-Gaussian data.
+The EPCA objective is formulated as a Bregman divergence [Bregman](@cite). Bregman divergences are a measure of difference between two points (often probability distributions); however, they are not proper metrics, because they do not always satisfy symmetry and the triangle inequality.
 
 ## Definition
 
-Formally, the Bregman divergence [Bregman](@cite) $B_F$ associated with a function $F(\theta)$ is defined as
+Formally, the Bregman divergence $B_F$ associated with a function $F(\theta)$ is defined as
 
 ```math
 B_F(p \| q) = F(p) - F(q) - \langle f(p), p - q \rangle
@@ -14,14 +12,14 @@ B_F(p \| q) = F(p) - F(q) - \langle f(p), p - q \rangle
 
 where 
 *  $F(\mu)$ is a strictly convex and continuously differentiable function, 
-*  $f(\mu) = \nabla_\mu F(\mu)$ is the convex conjugate (defined later) of $F$, 
+*  $f(\mu) = \nabla_\mu F(\mu)$ is the gradient of $F$, 
 *  and $\langle \cdot, \cdot \rangle$ denotes an inner product.
 
 Intuitively, the Bregman divergence expresses the difference at $p$ between $F$ and its first-order Taylor expansion about $q$.
 
-### Properties
+### Aside: Properties
 
-Unlike traditional metrics, Bregman divergences are not generally symmetric (i.e., $B_F(p \| q) \neq B_F(q \| p)$) and do not usually satsify the triangle inequality. However, they are always non-negative ($B_F(p \| q) \geq 0$) and equal $0$ if and only if $p = q$.
+Bregman divergences vanish $B_F(p \| q) = 0$ if and only if their inputs also vanish $p = q = 0$. They are also always non-negative $B_F(p \| q) \geq 0$ for all $p, q \in \mathrm{domain}(F)$.
 
 !!! info
     While the full EPCA objective is always non-negative, the `EPCA` loss may be negative because `ExpFamilyPCA.jl` uses transformed objectives that are equivalent but not equal to minimizing a sum of Bregman divergences.  
@@ -47,45 +45,43 @@ The log-partition function $G(\theta)$ ensures that the probability distribution
 
 ### Key Parameters
 
-1.  **Natural Parameter** ($\theta$): This parameter controls the distribution’s shape in its canonical form. For example, the natural parameter for the Poisson distribution is $\log \lambda$.
-2.  **Expectation Parameter** ($\mu$): This is the expected value of the sufficient statistic,[^1] computed as the mean of the data under the distribution. In exponential family distributions, it is related to the natural parameter through the gradient of the log-partition function:
+1. The **natural parameter** $\theta$ controls the distribution’s shape in its canonical form. For example, the natural parameter for the Poisson distribution is $\log \lambda$.
+2. The **expectation parameter** $\mu$ is the expected value of the sufficient statistic,[^1] computed as the mean of the data under the distribution. In exponential family distributions, it is related to the natural parameter through the **link function** $g$:
 
 ```math
 \mu = \mathbb{E}_{\theta}[X] = \nabla_\theta G(\theta) = g(\theta)
 ```
[…]
