### 2024-12-12 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
Here's the cleaned up documentation:

# Math

## Bregman Divergences
- [x] $f$ is indeed the ordinary gradient of $F$. Fixed the typo.
- [x] $\theta = f(\mu)$ is beautiful, but admittedly obtuse at first glance. The equality follows from symmetry in the dual spaces. Since $f$ inverts the link function $g$, applying $f$ to $\mu = g(\theta)$ yields $f(\mu) = \theta$. Added a remark/reference to the appendix section on why $f$ inverts $g$.
- [x] $f$ and $F$ (and also $g$ and $G$) are related to the exp family through convex conjugation and the Bregman loss. -- I've also rewritten the Bregman divergences section to hopefully be much clearer on all fronts.

## EPCA Objectives

- [x] change $\mu$ to $\mu_0$
- [x] the decompression function is indeed $g$. This can be tricky (I had to double check myself), but it's easiest to see from the objective that $X \approx g(\Theta)$, so it is the proper function to recover our decompressed data.
- [x] I actually debated including a similar table to the one you propose in the original documentation but ultimately decided against it because the parametric EPCA models use performance hacks that don’t fit neatly into categories like EPCA1 vs EPCA2. Creating a similar table for EPCA objectives 1–7 is also tricky since (1) they’re theoretically equivalent, and (2) multiple objectives can work for the same distribution. I prefer not to prescribe which distributions align with which objective. That said, if you feel strongly about it, I’m happy to create a table.

## Gamma Constructor

- [x] The gamma constructor was one of the trickiest to create b/c small instability in the solver kept yielding negative log errors. The solution was to widen the bound from theory (obvious in hindsight). The bound that worked for gammas was `-eps()`. I also created the `NegativeDomain` and `PositiveDomain` with more generous bounds so that people don't have to sorry about this for quick usage. My informal analysis also suggested that these bounds had a negligible impact on performance.

## API Documentation
- [x] Added an admonition clarifying the `metaprogramming` flag.
