### 2024-12-12 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
> It's a shame, codecov is very useful to figure out if some parts of your code are untested. Can you ask one of the admins to set it up.

I just requested privileges/help setting up the codecov badge. TBD on whether this is possible.

> Since we don't have codecov yet, I looked at the tests to understand what they cover. Why do you run a full test on some distributions and only a smoke test on others?

This is a very insightful question and it was the largest roadblock while developing the package. The reason that not all distributions are tested under all objectives is b/c of domain errors. Some distributions that have objective functions with restricted domains are unfeasible/unstable to optimize under certain specifications, so we don't test every specification. However, we still perform a baseline sanity test `smoke_test` to make sure everything performs well even if they can't be tested under every specification.
