+++ .github/workflows/CI.yml
@@ -48,12 +48,12 @@ jobs:
             ${{ runner.os }}-      
       - uses: julia-actions/julia-buildpkg@latest
       - uses: julia-actions/julia-runtest@latest
-      # - uses: julia-actions/julia-processcoverage@v1
-      # - uses: codecov/codecov-action@v4
-        # with:
-        #   files: lcov.info
-        #   token: ${{ secrets.CODECOV_TOKEN }}
-        #   fail_ci_if_error: false
+      - uses: julia-actions/julia-processcoverage@v1
+      - uses: codecov/codecov-action@v4
+        with:
+          files: lcov.info
+          token: ${{ secrets.CODECOV_TOKEN }}
+          fail_ci_if_error: false
   docs:
     name: Documentation
     runs-on: ubuntu-latest
