### 2024-12-16 98b75e2
added more tests to improve codecov
