+++ test/family/test_gamma.jl
@@ -1,7 +1,7 @@
 @testset "Gamma" begin
     n = 2
-    indim = 5
-    outdim = 5
+    indim = 3
+    outdim = 3
     X = rand(n, indim) * 100
 
     M1 = GammaEPCA(indim, outdim)
@@ -11,6 +11,6 @@
         M1,
         A1,
         X,
-        atol=2.5
+        atol=1.5
     )
 end
