### 2024-12-16 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
@gdalle, I've also just pushed another update to improve the codecov :) was a very good idea to set it up!
