+++ Project.toml
@@ -4,7 +4,6 @@ authors = ["Logan-Mondal-Bhamidipaty"]
 version = "2.0.0"
 
 [deps]
-CompressedBeliefMDPs = "0a809e47-b8eb-4578-b4e8-4c2c5f9f833c"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 FunctionWrappers = "069b7b12-0de2-55c6-9aab-29f3d0a68a2e"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
@@ -15,7 +14,6 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
 [compat]
-CompressedBeliefMDPs = "1"
 Distances = "0.10"
 FunctionWrappers = "1"
 LogExpFunctions = "0.3"
