### 2025-01-05 sisl/ExpFamilyPCA.jl#21 @FlyingWorkshop
@gdalle, I added a note to the README about

> > Regarding the README test: this is a great idea, but the snippets currently use sample_from_poisson which is a generic helper that is not implemented, so cannot test the method.
>
> Regardless of whether the README is tested during CI, it seems weird to put code in a code example that people can't run. Can you maybe replace `sample_from_poisson(...)` with `rand(Distributions.Poisson(lambda), ...)`?

I added a note to the README that should hopefully clear this up. I don't want to replace `sample_from_poisson` even if it is an unimplemented dummy function b/c I think it would harm code legibility.
