### 2025-01-05 #7403 @FlyingWorkshop
Hi @gdalle and @lrnv, happy new year! 🎉

I believe I've addressed all of points in the issue @gdalle opened (pending his approval). @gdalle could you please confirm everything has been resolved? If there are areas that still need improvement, I'll happily and promptly address them.

The second round of PhD applications is closing soon, and I’m hopeful we can have this published before then. Thank you!
