### 2025-01-06 #7403 @FlyingWorkshop
@gdalle, thank you so much for your feedback and for closing the review issue--it’s been incredibly helpful and deeply appreciated.

To hopefully make things easier, I’ve gone ahead and highlighted where I believe the remaining checklist items have been addressed. Of course, if there’s anything you feel still needs work or clarification, I’d be more than happy to make adjustments or discuss further.

### Functionality
- **Functionality:** From our [conversation about benchmarking](https://github.com/sisl/ExpFamilyPCA.jl/issues/21#issuecomment-2487019955), it seems the functionality claims are confirmed.
- **Performance:** I believe the performance claims have also been addressed, but I’m happy to add clarifying details or run additional tests if needed.

### Documentation
- **Statement of Need:** The paper (L17-35) provides a clear statement of need.
- **Community Guidelines:** Contribution guidelines are included in the [README](https://github.com/sisl/ExpFamilyPCA.jl?tab=readme-ov-file#contributing).

### Software Paper
- **State of the Field:** I remark on the state of the field in the statement of need (L18-26). In particular, I note that EPCA is the only package to support EPCA for multiple/custom distributions, and the only other package is a MATLAB API for Poisson EPCA. There are of course other packages for dimensionality reduction in Julia (e.g., MultivariateStats.jl, ManifoldLearning.jl), but they are sufficiently orthogonal that I believe omitting them is acceptable in the paper.
- **References:** To the best of my knowledge, the references are complete and properly cited.

Thank you again for your guidance and your patience with my responses--please don’t feel rushed on my account. I greatly appreciate your help and will address any additional feedback promptly.
