### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @gdalle
Example from JOSS paper lags forever with reasonable dimensions
The following example from the paper runs in very small dimension, but even for moderately-sized settings it lags forever (> 1 min). I tested it on a very bulky Macbook Pro so I really don't think the problem comes from the hardware. Do you have any clue why it is so slow?

```julia
using ExpFamilyPCA, Distributions

indim, outdim, n1 = 10, 5, 100

sample_from_gamma(n, m) = rand(Gamma(), n, m)

G(θ) = -log(-θ)
g(θ) = -1 / θ
gamma_epca = EPCA(indim, outdim, G, g, Val((:G, :g)); options = NegativeDomain())

X = sample_from_gamma(n1, indim)  # matrix of gamma-distributed data

X_compressed = fit!(gamma_epca, X)
```
