### 2025-01-06 sisl/ExpFamilyPCA.jl#26 @gdalle
Reproducibility of plots in JOSS paper
To ensure complete reproducibility of the plots in the paper, can you please:

- specify the complete environment where the experiments are run, with a project and a manifest
- explain how the JSON files `pca_data.json` and `epca_data.json` are obtained
- create the random generator with StableRNGs.jl to fix it across Julia versions
