### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @gdalle
My bad, I should have switched on verbose mode. I forgot that we were solving a possibly challenging optimization problem. Perhaps a warning would be appropriate, telling users that the default `fit!` may be long and that they can play with verbosity or iteration number?
