### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @FlyingWorkshop
No worries, very valid concern. I can add a disclaimer to the documentation now.
