### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @gdalle
Also, if you add a `yield()` at the end of every iteration (or every couple of iterations), users will be able to interrupt optimization without spamming `ctrl+c` and eventually crashing their REPL
