### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @gdalle
Try it, it's life-changing
