### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @FlyingWorkshop
Will look into `yield` now too
