### 2025-01-06 #7403 @gdalle
The remaining issues for me are:

- https://github.com/sisl/ExpFamilyPCA.jl/issues/25
- https://github.com/sisl/ExpFamilyPCA.jl/issues/26
- https://github.com/sisl/ExpFamilyPCA.jl/issues/27
- the missing DOI

After that, the paper will be good to go
