### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @FlyingWorkshop
I added the yields. I'm not familiar w/ this mechanic of Julia, so I'm not sure if this accomplishes the desired result, but it does seem to let me ctrl+c in the middle of `fit!` w/o crashing the repo, which is the desired behavior.
