### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @FlyingWorkshop
What do you think @gdalle?
