### 2025-01-06 sisl/ExpFamilyPCA.jl#25 @gdalle
Looks good! Let's close this one
