### 2025-01-06 bfaac2d
expanded plot reproducability
