+++ scripts/Project.toml
@@ -1,0 +1,9 @@
+[deps]
+CompressedBeliefMDPs = "0a809e47-b8eb-4578-b4e8-4c2c5f9f833c"
+ExpFamilyPCA = "9c724b78-6801-4402-8a63-53f028696012"
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+POMDPTools = "7588e00f-9cae-40de-98dc-e0c70c48cdd7"
+POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
