### 2025-01-06 sisl/ExpFamilyPCA.jl#26 @FlyingWorkshop
Done :)

Changes:

- made project/manifest in scripts folder
- the jsons are obtained from this [script](https://github.com/sisl/ExpFamilyPCA.jl/blob/main/scripts/belief_compression.jl)
- RNG uses `StableRNG`
- minor parameter tweaks to make graphs look nice w/ new seed
