### 2025-01-06 sisl/ExpFamilyPCA.jl#26 @FlyingWorkshop
Let me know what you think @gdalle
