### 2025-01-06 sisl/ExpFamilyPCA.jl#26 @gdalle
if you can also turn the jupyter notebook into a script + integrate the new plots into the paper, I'll test it again tomorrow and then we'll be able to close this
