### 2025-01-06 245fb69
converted notebook to script
