### 2025-01-06 sisl/ExpFamilyPCA.jl#26 @FlyingWorkshop
Here is a [script](https://github.com/sisl/ExpFamilyPCA.jl/blob/main/scripts/belief_script.jl). The new plot is in the paper too.
