### 2025-01-07 sisl/ExpFamilyPCA.jl#26 @gdalle
The script doesn't seem to run because `x` is not defined on line 33. Moreover, you should probably use relative paths with e.g. `@__DIR__` so that the saving location doesn't depend on where the REPL was launched.
