### 2025-01-07 sisl/ExpFamilyPCA.jl#26 @FlyingWorkshop
That's so strange, it ran fine when I was running it (perhaps I accidentally ran it twice and didn't notice it failed the first time, so that `x` was defined in the REPL the second run). I've updated the script and path dependencies so that it should run now (I just removed the path prefixes).

The script should create several images (mirroring the original notebook), but only the last image `combo.png` matters for the paper.
