### 2025-01-07 62ed5e7
major plotting script cleanup
