+++ scripts/make_plots.jl
@@ -1,0 +1,186 @@
+using ExpFamilyPCA
+using CompressedBeliefMDPs
+using POMDPs, POMDPTools
+using StableRNGs
+using JSON
+using Plots
+using Statistics
+
+const MAX_OUTDIMS = 6
+const LOGO_COLORS = Colors.JULIA_LOGO_COLORS
+
+function kl_div(p, q; ϵ=eps())
+    return sum(@. p * log((p + ϵ) / (q + ϵ)))
+end
+
+function to_probability(matrix::AbstractMatrix)
+    matrix = clamp.(matrix, 0, 1)
+    prob_matrix = matrix ./ sum(matrix, dims=2)
+    return prob_matrix
+end
+
+function calc_kl(
+    epca::EPCA, 
+    X::AbstractMatrix{T}
+) where T<:Real
+    A = ExpFamilyPCA.fit!(epca, X; maxiter=100, verbose=true, steps_per_print=10)
+    X_recon = decompress(epca, A)
+    Q = to_probability(X_recon)
+    divergences = []
+    for (p, q) in zip(eachrow(X), eachrow(Q))
+        kl = kl_div(p, q)
+        push!(divergences, kl)
+    end
+    result = mean(divergences)  # Return the mean KL divergence
+    @show result
+    return result
+end
+
+
+function make_figure1()
+    n_corridors = 2
+    corridor_length = 20
+    maze = CircularMaze(n_corridors, corridor_length)
+
+    rng = StableRNG(100)
+    policy = RandomPolicy(maze; rng=rng)
+    sampler = PolicySampler(maze; policy=policy, rng=rng, n=200)
+
+    raw_beliefs = sampler(maze)
+    beliefs = make_numerical(raw_beliefs, maze)
+
+    _, indim = size(beliefs)
+
+    outdims = 1:MAX_OUTDIMS
+
+    kl_divs_poisson_epca = []
+    kl_divs_gaussian_epca = []
+
+    for outdim in outdims
+        @show outdim
+        epca_kl = calc_kl(PoissonEPCA(indim, outdim), beliefs)
+        pca_kl = calc_kl(GaussianEPCA(indim, outdim), beliefs)
+        push!(kl_divs_poisson_epca, epca_kl)
+        push!(kl_divs_gaussian_epca, pca_kl)
+    end
+
+    plot(
+        outdims, 
+        kl_divs_poisson_epca, 
+        label="EPCA", 
+        color=LOGO_COLORS.blue,
+        yscale=:log10,
+        marker=:cross, 
+        linestyle=:solid, 
+        lw=2,
+        dpi=600
+    )
+    plot!(
+        outdims, 
+        kl_divs_gaussian_epca, 
+        label="PCA",
+        yscale=:log10,
+        color=LOGO_COLORS.green,
+        marker=:x,
+        linestyle=:dash,
+        lw=2,
+        dpi=600
+
+    )
+    label_size = 8
+    tick_size = 8
+    title!("KL Divergence Across Bases", fontsize=14)
+    xlabel!("Number of Bases", fontsize=label_size)
+    ylabel!("Average KL Divergence", fontsize=label_size)
+    figure1 = plot!(
+        legendfontsize = 10,
+        xtickfontsize = tick_size,
+        ytickfontsize = tick_size,
+        framestyle = :box,
+        aspectration = :equal,
+        markerstrokewidth = 2,  # Makes markers more prominent
+        markeralpha = 0.7  # Adds transparency to markers
+    )
+
+    return figure1
+end
+
+function is_terminal_belief(belief)
[…]
