### 2025-01-07 sisl/ExpFamilyPCA.jl#26 @FlyingWorkshop
@gdalle, I cleaned up the plotting script significantly. The whole pipeline now runs from a [single file](https://github.com/sisl/ExpFamilyPCA.jl/blob/main/scripts/make_plots.jl) and there are no extraneous images/JSONs. This should hopefully be very easy to test.
