### 2025-01-09 #7403 @lrnv
@FlyingWorkshop can you take care of your part of the [post review checklist](https://github.com/openjournals/joss-reviews/issues/7403#issuecomment-2536150505) and report back to me ?
