+++ .github/workflows/CI.yml
@@ -54,6 +54,7 @@ jobs:
           files: lcov.info
           token: ${{ secrets.CODECOV_TOKEN }}
           fail_ci_if_error: false
+          verbose: true
   docs:
     name: Documentation
     runs-on: ubuntu-latest
