+++ v2.0.2/refs.bib
@@ -1,0 +1,171 @@
+@article{EPCA,
+  title={A Generalization of Principal Components Analysis to the Exponential Family},
+  author={Collins, Michael and Dasgupta, Sanjoy and Schapire, Robert E},
+  journal={Advances in Neural Information Processing Systems},
+  volume={14},
+  year={2001}
+}
+
+@article{azoury,
+  title={Relative Loss Bounds for On-Line Density Estimation with the Exponential Family of Distributions},
+  author={Azoury, Katy S and Warmuth, Manfred K},
+  journal={Machine learning},
+  volume={43},
+  pages={211--246},
+  year={2001},
+  publisher={Springer}
+}
+
+@book{elements,
+  title={The {E}lements of {S}tatistical {L}earning: {D}ata {M}ining, {I}nference, and {P}rediction},
+  author={Hastie, Trevor and Tibshirani, Robert and Friedman, Jerome H and Friedman, Jerome H},
+  volume={2},
+  year={2009},
+  publisher={Springer},
+  doi={10.1007/978-0-387-84858-7}
+}
+
+@article{composition,
+  title={Compositional data analysis},
+  author={Greenacre, Michael},
+  journal={Annual Review of Statistics and its Application},
+  volume={8},
+  number={1},
+  pages={271--299},
+  year={2021},
+  publisher={Annual Reviews}
+}
+
+@misc{dispatch,
+  author       = {Stefan Karpinski},
+  title        = {The Unreasonable Effectiveness of Multiple Dispatch},
+  year         = {2019},
+  howpublished = {Conference Talk at JuliaCon 2019, available at \url{https://www.youtube.com/watch?v=kc9HwsxE1OY}},
+  note         = {Accessed: 2024-09-13}
+}
+
+@article{symbolics,
+author = {Gowda, Shashi and Ma, Yingbo and Cheli, Alessandro and Gw\'{o}\'{z}zd\'{z}, Maja and Shah, Viral B. and Edelman, Alan and Rackauckas, Christopher},
+title = {High-Performance Symbolic-Numerics via Multiple Dispatch},
+year = {2022},
+issue_date = {September 2021},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {55},
+number = {3},
+issn = {1932-2240},
+url = {https://doi.org/10.1145/3511528.3511535},
+doi = {10.1145/3511528.3511535},
[…]
