### 2025-01-09 a18ac15
Update Project.toml
For zendozo archive
