+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ExpFamilyPCA"
 uuid = "9c724b78-6801-4402-8a63-53f028696012"
 authors = ["Logan-Mondal-Bhamidipaty"]
-version = "2.0.2"
+version = "2.0.3"
 
 [deps]
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
