### 2025-01-10 #7403 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
