### 2025-01-14 #7403 @crvernon
🥳 Congratulations on your new publication @FlyingWorkshop! Many thanks to @lrnv for editing and @ManuelStapper, @gdalle and @dufourc1 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
